renaming SSR file to containerized deployment|revamped sidebar code with necessary chnages
