+++ packages/manager/src/views/WebPropertyDetailPage/components/SSR/ContainerizedDeployment.tsx
